+++ docs/changelog.rst
@@ -34,6 +34,7 @@ New Features
 - Created new tests to cover the various calculation methods in :class:`~gatoh.model.ABModel`
 - QoL method :meth:`~gatoh.model.ABModel.add_relationships_to_hierarchy`
 - Defined multiple global constants throughout modules to clarify "magic numbers"
+- Prototype for ability to add "link functions" to model parameters
 
 Fixes
 -----